+++ app.go
@@ -15,11 +15,12 @@ var defaults []ApplicationOption
 
 // Application collects all dependencies and exposes them in a single context.
 type Application struct {
-	logger       log.Logger
-	errorHandler emperror.Handler
-	tracer       opentracing.Tracer
-	closers      []io.Closer
-	entries      map[string]interface{}
+	logger         log.Logger
+	errorHandler   emperror.Handler
+	tracer         opentracing.Tracer
+	closers        []io.Closer
+	entries        map[string]interface{}
+	lifecycleHooks []Hook
 }
 
 func NewApplication(opts ...ApplicationOption) *Application {