+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 
 - Add arbitrary entries to the application
+- Lifecycle management
 
 
 ## 0.2.0 - 2017-10-02